+++ packages/dropdown_button2/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix memory leak in CurvedAnimation [Flutter core].
 - Avoid Container objects when possible for better performance [Flutter core].
 - Add semantics to dropdown menu items [Flutter core].
+- Support helperStyle/helperMaxLines/errorMaxLines for DropdownButtonFormField2.
 
 ## 3.0.0-beta.21